move card styles to commun
