+++ src/partner/modules/NavBar/NavBar.tsx
@@ -13,20 +13,20 @@ const NavBarStyle = {
   fontFamily: "unset",
   fontSize: "14px",
   textDecoration: "none",
-  color: "Gray"
+  color: "Gray",
 };
 
 const NavBarActiveStyle = {
   ...NavBarStyle,
   color: "#E83E5D",
   background: "WhiteSmoke",
-  fontStyle: "bold"
+  fontStyle: "bold",
 };
 
 export const Img = styledComponents.img({
   width: "2.5em",
   height: "2.5em",
-  objectFit: "contain"
+  objectFit: "contain",
 });
 
 interface INavBarProps {
