+++ src/partner/modules/ui/index.ts
@@ -3,19 +3,6 @@ import * as ListStyled from "./List/List";
 import { Modal } from "./Modal";
 import * as Notification from "./Modal/Notification";
 
-export {
-  CardContainer,
-  CardRowHeader,
-  CardTextHeaderContainer,
-  CardRow,
-  CardDescription,
-  CardDivActionsContainer,
-  CardIconImg,
-  CardLinkIcon,
-  CardOrderSection,
-  RowProducts,
-} from "./Card";
-
 export * from "./Header";
 
 export { NavBarStyled, ListStyled, Modal, Notification };
